Move OS specific library function implementations to libzutil
