+++ cmd/zfs/Makefile.am
@@ -18,4 +18,5 @@ zfs_LDADD = \
 	$(top_builddir)/lib/libnvpair/libnvpair.la \
 	$(top_builddir)/lib/libuutil/libuutil.la \
 	$(top_builddir)/lib/libzfs/libzfs.la \
+	$(top_builddir)/lib/libzutil/libzutil.la \
 	$(top_builddir)/lib/libzfs_core/libzfs_core.la
