+++ cmd/zinject/Makefile.am
@@ -13,4 +13,5 @@ zinject_SOURCES = \
 
 zinject_LDADD = \
 	$(top_builddir)/lib/libnvpair/libnvpair.la \
-	$(top_builddir)/lib/libzfs/libzfs.la
+	$(top_builddir)/lib/libzfs/libzfs.la \
+	$(top_builddir)/lib/libzutil/libzutil.la
