+++ cmd/zpool/Makefile.am
@@ -16,6 +16,7 @@ zpool_SOURCES = \
 zpool_LDADD = \
 	$(top_builddir)/lib/libnvpair/libnvpair.la \
 	$(top_builddir)/lib/libuutil/libuutil.la \
+	$(top_builddir)/lib/libzutil/libzutil.la \
 	$(top_builddir)/lib/libzfs/libzfs.la
 
 zpool_LDADD += -lm $(LIBBLKID)
