+++ lib/libspl/Makefile.am
@@ -20,7 +20,6 @@ noinst_LTLIBRARIES = libspl.la
 USER_C = \
 	getexecname.c \
 	gethostid.c \
-	getmntany.c \
 	list.c \
 	mkdirp.c \
 	page.c \
