+++ lib/libspl/include/sys/mnttab.h
@@ -67,10 +67,13 @@ struct extmnttab {
 	uint_t mnt_minor;
 };
 
+struct stat64;
+struct statfs;
+
 extern int getmntany(FILE *fp, struct mnttab *mp, struct mnttab *mpref);
 extern int _sol_getmntent(FILE *fp, struct mnttab *mp);
-extern int getextmntent(FILE *fp, struct extmnttab *mp, int len);
-
+extern int getextmntent(const char *path, struct extmnttab *mp,
+    struct stat64 *statbuf);
 static inline char *_sol_hasmntopt(struct mnttab *mnt, char *opt)
 {
 	struct mntent mnt_new;
